Fixing integer and integer ranges
This has a knock on effect to min_max_repetition as previously 'nothing'
would match the rule for an integer, so repetition_min would return the
empty string rather than nil.  The logic rationalising the repetitions
needs to be changed to reflect this.  I tried to do that but failed!