removing build for 2.0.0
The latest version of Addressable (2.5.2) fixes a number of important bugs but also now use public_suffix 3, which won't work on Ruby 2.0 any more. Therefore, our builds fail with 2.0. Ruby 2.0 is out of maintenance anyway, so I'm removing support for it.
I guess anybody on an older system can use rvm or rbenv to get a newer build of Ruby.